AUT-4766: Rename files to be kebab case
