+++ journey-map/src/helpers/hex-helper.ts
